Adapt to use GCHP outputs

+++ src/components/template_component/template.sh
@@ -9,7 +9,11 @@
 setup_template() {
     printf "\n=== CREATING TEMPLATE RUN DIRECTORY ===\n"
 
-    cd ${GCClassicPath}/run
+    if "$UseGCHP"; then
+        cd ${GCHPPath}/run
+    else
+        cd ${GCClassicPath}/run
+    fi
 
     # The createRunDir.sh script assumes the file ~/.geoschem/config exists
     # and contains the path to GEOS-Chem input data
@@ -36,35 +40,45 @@ setup_template() {
         printf "\n Options are GEOSFP or MERRA2.\n"
         exit 1
     fi
-    if [ "$Res" = "4.0x5.0" ]; then
-        cmd="9\n${metNum}\n1\n2\n${RunDirs}\n${runDir}\nn\n"
-    elif [ "$Res" == "2.0x2.5" ]; then
-        cmd="9\n${metNum}\n2\n2\n${RunDirs}\n${runDir}\nn\n"
-    elif [ "$Res" == "0.5x0.625" ]; then
-        if "$isRegional"; then
-            # Use NA domain by default and adjust lat/lon below
-            cmd="9\n${metNum}\n3\n4\n2\n${RunDirs}\n${runDir}\nn\n"
+
+    if "$UseGCHP"; then
+        if [ "${metNum}" == "1"]; then
+            cmd="5\n2\n${metNum}\n${RunDirs}\n${runDir}\nn\n"
         else
-            cmd="9\n${metNum}\n3\n1\n2\n${RunDirs}\n${runDir}\nn\n"
-        fi
-    elif [ "$Res" == "0.25x0.3125" ]; then
-        if "$isRegional"; then
-            # Use NA domain by default and adjust lat/lon below
-            cmd="9\n${metNum}\n4\n4\n2\n${RunDirs}\n${runDir}\nn\n"
-        else
-            cmd="9\n${metNum}\n4\n1\n2\n${RunDirs}\n${runDir}\nn\n"
-        fi
-    elif [ "$Res" == "0.125x0.15625" ]; then
-        if "$isRegional"; then
-            # Use NA domain by default and adjust lat/lon below
-            cmd="9\n${metNum}\n5\n4\n2\n${RunDirs}\n${runDir}\nn\n" #regional run
-        else
-            cmd="9\n${metNum}\n5\n1\n2\n${RunDirs}\n${runDir}\nn\n"
+            # GEOSFP: Use daily files pre-processed for GEOS-Chem
+            cmd="5\n2\n${metNum}\n1\n${RunDirs}\n${runDir}\nn\n"
         fi
     else
-        printf "\nERROR: Grid resolution ${Res} is not supported by the IMI. "
-        printf "\n Options are 0.125x0.15625, 0.25x0.3125, 0.5x0.625, 2.0x2.5, or 4.0x5.0.\n"
-        exit 1
+        if [ "$Res" = "4.0x5.0" ]; then
+            cmd="9\n${metNum}\n1\n2\n${RunDirs}\n${runDir}\nn\n"
+        elif [ "$Res" == "2.0x2.5" ]; then
+            cmd="9\n${metNum}\n2\n2\n${RunDirs}\n${runDir}\nn\n"
+        elif [ "$Res" == "0.5x0.625" ]; then
+            if "$isRegional"; then
+                # Use NA domain by default and adjust lat/lon below
+                cmd="9\n${metNum}\n3\n4\n2\n${RunDirs}\n${runDir}\nn\n"
+            else
+                cmd="9\n${metNum}\n3\n1\n2\n${RunDirs}\n${runDir}\nn\n"
+            fi
+        elif [ "$Res" == "0.25x0.3125" ]; then
+            if "$isRegional"; then
+                # Use NA domain by default and adjust lat/lon below
+                cmd="9\n${metNum}\n4\n4\n2\n${RunDirs}\n${runDir}\nn\n"
+            else
+                cmd="9\n${metNum}\n4\n1\n2\n${RunDirs}\n${runDir}\nn\n"
+            fi
+        elif [ "$Res" == "0.125x0.15625" ]; then
+            if "$isRegional"; then
+                # Use NA domain by default and adjust lat/lon below
+                cmd="9\n${metNum}\n5\n4\n2\n${RunDirs}\n${runDir}\nn\n" #regional run
+            else
+                cmd="9\n${metNum}\n5\n1\n2\n${RunDirs}\n${runDir}\nn\n"
+            fi
+        else
+            printf "\nERROR: Grid resolution ${Res} is not supported by the IMI. "
+            printf "\n Options are 0.125x0.15625, 0.25x0.3125, 0.5x0.625, 2.0x2.5, or 4.0x5.0.\n"
+            exit 1
+        fi
     fi
 
     # Create run directory
@@ -77,9 +91,22 @@ setup_template() {
     # Copy download script to run directory
     cp ${InversionPath}/src/utilities/download_gc_data.py download_gc_data.py
 
-    # Modify geoschem_config.yml based on settings in config.yml
-    sed -i -e "s:20190101:${StartDate}:g" \
-        -e "s:20190201:${EndDate}:g" geoschem_config.yml
+    if "$UseGCHP"; then
+        sed -i -e "s:20190101:${StartDate}:g" cap_restart
+        sed -i -e "s/Run_Duration=\"[0-9]{8} 000000\"/Run_Duration=\"${RunDuration} 000000\"/" \
+            -e 's/^CS_RES=.*$/CS_RES="${CS_RES}"/' \
+            -e 's/^TOTAL_CORES=.*$/TOTAL_CORES="${TOTAL_CORES}"/' \
+            -e 's/^NUM_NODES=.*$/NUM_NODES="${NUM_NODES}"/' \ 
+            -e 's/^NUM_CORES_PER_NODE=.*$/NUM_CORES_PER_NODE="${NUM_CORES_PER_NODE}"/' \
+            -e 's/^AutoUpdate_Diagnostics=.*$/AutoUpdate_Diagnostics=OFF/' \
+            -e 's/^Diag_Monthly=.*$/Diag_Monthly="0"/' \
+            -e 's/^Diag_Frequency=.*$/Diag_Frequency="240000"/' \
+            -e 's/^Diag_Duration=.*$/Diag_Duration="240000"/' setComminRunSettings.sh
+    else
+        # Modify geoschem_config.yml based on settings in config.yml
+        sed -i -e "s:20190101:${StartDate}:g" \
+            -e "s:20190201:${EndDate}:g" geoschem_config.yml
+    fi
 
     if "$isRegional"; then
         # Adjust lat/lon bounds because GEOS-Chem defines the domain
@@ -97,10 +124,16 @@ setup_template() {
     OLD=" StateVector.nc"
     NEW=" ${RunDirs}/StateVector.nc"
     sed -i -e "s@$OLD@$NEW@g" HEMCO_Config.rc
+    if "$UseGCHP"; then
+        sed -i -e "s@$OLD@$NEW@g" ExtData.rc
+    fi
 
     # Modify HEMCO_Config.rc if running Kalman filter
     if "$KalmanMode"; then
         sed -i -e "s|gridded_posterior.nc|${RunDirs}/ScaleFactors.nc|g" HEMCO_Config.rc
+        if "$UseGCHP"; then
+            sed -i -e "s|gridded_posterior.nc|${RunDirs}/ScaleFactors.nc|g" ExtData.rc
+        fi
     fi
 
     # Modify HEMCO_Config.rc based on settings in config.yml
@@ -121,7 +154,10 @@ setup_template() {
 
     # By default, only output emissions at the end of the simulation
     sed -i -e "s|DiagnFreq:                   Monthly|DiagnFreq:                   End|g" HEMCO_Config.rc
-
+    # do not output Emissions collection
+    if "$UseGCHP"; then
+        sed -i "s/'Emissions',/#'Emissions',/" HISTORY.rc
+    fi
     # Add a new ZERO scale factor for use in jacobian simulations
     sed -i -e "/1 NEGATIVE       -1.0 - - - xy 1 1/a 5 ZERO            0.0 - - - xy 1 1" HEMCO_Config.rc
 
@@ -134,6 +170,9 @@ setup_template() {
     # Temporary fix: Modify path to HEMCO prior emissions (the path is currently
     # hardcoded in the template HEMCO config file in GEOS-Chem)
     sed -i -e "s|prior_run|hemco_prior_emis|g" HEMCO_Config.rc
+    if "$UseGCHP"; then
+        sed -i -e "s|prior_run|hemco_prior_emis|g" ExtData.rc
+    fi
     
     # Modify HISTORY.rc - comment out diagnostics that aren't needed
     sed -i -e "s:'CH4':#'CH4':g" \
@@ -147,37 +186,63 @@ setup_template() {
         -e "s:'Met_PFLLSAN:#'Met_PFLLSAN:g" HISTORY.rc
 
     # If turned on, save out hourly CH4 concentrations to daily files
+    # use time-average mode
     if "$HourlyCH4"; then
         sed -i -e 's/SpeciesConc.frequency:      00000100 000000/SpeciesConc.frequency:      00000000 010000/g' \
-            -e 's/SpeciesConc.duration:       00000100 000000/SpeciesConc.duration:       00000001 000000/g' \
-            -e 's/SpeciesConc.mode:           '\''time-averaged/SpeciesConc.mode:           '\''instantaneous/g' HISTORY.rc
+            -e 's/SpeciesConc.duration:       00000100 000000/SpeciesConc.duration:       00000001 000000/g' HISTORY.rc
     fi
 
-    # Remove sample restart file
-    rm -f Restarts/GEOSChem.Restart.20190101_0000z.nc4
+    # Remove sample restart file; GCHP restarts are just soft links and are needed later as template
+    if [ "$UseGCHP" != true ]; then
+        rm -f Restarts/GEOSChem.Restart.20190101_0000z.nc4
+    fi
 
     # Copy template run script
-    cp ${InversionPath}/src/geoschem_run_scripts/ch4_run.template .
+    if "$UseGCHP"; then
+        cp ${InversionPath}/src/geoschem_run_scripts/gchp_ch4_run.template .
+    else
+        cp ${InversionPath}/src/geoschem_run_scripts/ch4_run.template .
+    fi
 
     # Copy input file for applying emissions perturbations via HEMCO
-    cp ${InversionPath}/src/geoschem_run_scripts/Perturbations.txt .
+    if [ "$UseGCHP" != true ]; then
+        cp ${InversionPath}/src/geoschem_run_scripts/Perturbations.txt .
+    fi
 
     # Compile GEOS-Chem and store executable in GEOSChem_build directory
-    printf "\nCompiling GEOS-Chem...\n"
     cd build
-    cmake ${InversionPath}/GCClassic >>build_geoschem.log 2>&1
-    cmake . -DRUNDIR=.. >>build_geoschem.log 2>&1
+    if "$UseGCHP"; then
+        printf "\nCompiling GCHP...\n"
+        cmake ${InversionPath}/GCHP >>build_geoschem.log 2>&1
+    else
+        printf "\nCompiling GEOS-Chem...\n"
+        cmake ${InversionPath}/GCClassic >>build_geoschem.log 2>&1
+    fi
+    cmake . -DRUNDIR=.. -DMECH=carbon >>build_geoschem.log 2>&1
     make -j install >>build_geoschem.log 2>&1
     cd ..
-    if [[ -f gcclassic ]]; then
-        rm -rf build
-        mv build_info ../GEOSChem_build
-        mv -v gcclassic ../GEOSChem_build/
+    if "$UseGCHP"; then
+        if [[ -f gchp ]]; then
+            mkdir ../GEOSChem_build
+            mv -v gchp ../GEOSChem_build/
+            mv build/CMakeCache.txt ../GEOSChem_build
+            mv build/build_geoschem.log ../GEOSChem_build
+            rm -rf build
+        else
+            printf "\nGCHP build failed! \n\nSee ${RunTemplate}/GEOSChem_build/build_geoschem.log for details\n"
+            exit 999
+        fi
     else
-        printf "\nGEOS-Chem build failed! \n\nSee ${RunTemplate}/build/build_geoschem.log for details\n"
-        exit 999
-    fi
-    printf "\nDone compiling GEOS-Chem \n\nSee ${RunDirs}/GEOSChem_build_info for details\n\n"
+        if [[ -f gcclassic ]]; then
+            mv build_info ../GEOSChem_build
+            mv -v gcclassic ../GEOSChem_build/
+            mv build/build_geoschem.log ../GEOSChem_build
+            rm -rf build
+        else
+            printf "\nGEOS-Chem build failed! \n\nSee ${RunTemplate}/build/build_geoschem.log for details\n"
+            exit 999
+        fi
+    printf "\nDone compiling GEOS-Chem \n\nSee ${RunDirs}/GEOSChem_build for details\n\n"
 
     # Navigate back to top-level directory
     cd ..
